Collapse regnode EXACTFU_SS into EXACTFUP
EXACTFUP was created by the previous commit to handle a problematic case
in which not all the code points in an EXACTFU node are /i foldable at
compile time.  Doing so will allow a future commit to use the pre-fold
EXACTFU nodes (done in a prior commit), saving execution time for the
common case.  The only problematic code point is the MICRO SIGN.  Most
patterns don't use this character.

EXACTFU_SS is problematic in a different way.  It contains the sequence
'ss' which is folded to by LATIN SMALL LETTER SHARP S, but everything in
it can be pre-folded (unless it also contains a MICRO SIGN).  The reason
this is problematic is that it is the only non-UTF-8 node where the
length in folding can change.  To process it at runtime, the more
general fold equivalence function is used that is capable of handling
length disparities, but is slower than the functions otherwise used for
non-UTF-8.

What I've chosen to do for now is to make a single node type for all the
problematic cases (which at this time means just the two aforementioned
ones).  If we didn't do this, we'd have to create a node type for
patterns that contain both 'ss' and MICRO.  Or artificially split the
pattern so the two never were in the same node.  If more special
handling is found to be needed or better, there'd be a combinatorial
explosion of nodes to handle all possible combinations.

What this effectively means is that the slower, more general foldEQ
function is used for the MICRO sign when the pattern isn't in UTF-*,
whereas there is no real underlying reason to do so for non-UTF-8
strings.