regen/mk_invlists.pl: Rmv outdated code
Before the GCB property handling got more complicated, it was possible
to represent it's vagaries with a boolean table on early Unicode
releases.  Now there are more complicated rules, and even though early
releases only use 0 or 1, the rules exist and lead to compilation
errors.  Just remove the special handling, and let the table be U8.