mktables: Make Turkic 'I' chars problematic
In a Turkic locale, these are problematic because their mappings
cross the 255/256 boundary.

This change has the side effect of causing U+307 to be added to the
problematic list, and it normally really isn't problematic, because in
those locales where U+130 and U+131 are problematic, U+307 isn't used.
But applications could switch in and out of Turkic locales, so it's best
to leave it be considered problematic.  The consequences of making this
mark problematic are simply slightly less optimized regex pattern code.